bin/requirements.txt
PyYAML==5.1.2
SQLAlchemy==1.3.4
asn1crypto==0.24.0
awscli==1.16.292
boto3==1.10.28
botocore==1.13.28
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
dateparser==0.7.1
docutils==0.15.2
future==0.17.1
humanize==0.5.1
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
pyasn1==0.4.8
pycparser==2.19
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
regex==2019.6.8
requests-oauthlib==1.2.0
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
schedule==0.6.0
six==1.13.0
twython==3.7.0
tzlocal==1.5.1
urllib3==1.25.7
